+++ module/Batch/src/Command/WithdrawStats/WithdrawStatsFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Batch\Command\WithdrawStats;
+
+class WithdrawStatsFactory
+{
+    public function __invoke($controllers)
+    {
+        return new WithdrawStats($controllers->get('faucetdev'));
+    }
+}