+++ module/Batch/src/Command/WthCurrencyStats/WthCurrencyStatsFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Batch\Command\WthCurrencyStats;
+
+class WthCurrencyStatsFactory
+{
+    public function __invoke($controllers)
+    {
+        return new WthCurrencyStats($controllers->get('faucetdev'));
+    }
+}